[refactor] Extract AudioData module
